pyproject.toml
[tool.poetry]
name = "mfti-project-team-21"
version = "0.1.0"
description = ""
authors = ["Chumakov Mikhail <chumakov-md@zvuk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
notebook = "^7.0.6"
pandas = "^2.1.4"
numpy = "^1.26.2"
jupyter = "^1.0.0"
openpyxl = "^3.1.2"
scikit-learn = "^1.3.2"
streamlit = "^1.29.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
